requirements.txt
Flask
Flask-Cors
langchain
langchain-community
langchain-core
langchain-google-genai
langchain-mongodb
langchain-pinecone
langchain-text-splitters
google-generativeai
google-auth
google-auth-httplib2
grpcio
grpcio-status
gunicorn
pypdf
neo4j
requests
APScheduler
chroma
chromadb
langchain-chroma
protobuf==3.20.*
